requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
apispec==6.0.2
apispec-webframeworks==0.5.2
attrs==22.2.0
bcrypt==4.2.1
build==1.0.3
CacheControl==0.13.1
cachelib==0.9.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.1.0
click==7.1.2
click-default-group==1.2.4
crashtest==0.4.1
cryptography==39.0.1
decorator==5.1.1
distlib==0.3.7
dnspython==2.6.1
dulwich==0.21.6
email_validator==2.2.0
exceptiongroup==1.1.0
fastapi==0.115.4
fastjsonschema==2.19.0
filelock==3.13.1
firestore-model==0.0.5
Flask==1.1.4
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Session==0.3.2
flask-session-plus==0.6.2
flask-swagger-ui==4.11.1
google-api-core==1.27.0
google-auth==1.24.0
google-cloud-bigquery==2.18.0
google-cloud-core==1.7.3
google-cloud-firestore==2.3.4
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.41.1
google-crc32c==1.5.0
google-resumable-media==1.3.3
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==1.1.0
jaraco.classes==3.3.0
Jinja2==2.11.3
joblib==1.2.0
jsonschema==4.17.3
jsonschema-spec==0.1.3
keyring==24.3.0
lazy-object-proxy==1.9.0
libcst==0.4.9
MarkupSafe==2.0.1
marshmallow==3.19.0
more-itertools==10.1.0
motor==3.6.1
ms-identity-web==0.16.6
msal==1.21.0
msgpack==1.0.7
mypy-extensions==1.0.0
numpy==1.24.2
openapi-schema-validator==0.4.3
openapi-spec-validator==0.5.5
packaging==23.0
passlib==1.7.4
pathable==0.4.3
pexpect==4.8.0
Pillow==9.1.1
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.0.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prettytable==3.6.0
proto-plus==1.22.2
protobuf==4.22.0
ptyprocess==0.7.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.6.0
pymongo==4.9.2
pyOpenSSL==19.1.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==3.5.2
requests==2.28.2
requests-toolbelt==0.9.1
retry==0.9.2
rfc3339-validator==0.1.4
rsa==4.9
schedule==1.1.0
scikit-learn==1.2.1
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlite-fts4==1.0.3
sqlite-utils==3.37
starlette==0.41.2
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2023.11.22
typing-inspect==0.8.0
typing_extensions==4.12.2
urllib3==1.26.14
uvicorn==0.32.0
uvloop==0.21.0
uWSGI==2.0.21
virtualenv==20.24.7
watchfiles==0.24.0
wcwidth==0.2.6
websockets==13.1
Werkzeug==1.0.1
xattr==0.10.1
zipp==3.15.0
